Finalized bTree ADT
Now ADT table does not show hash on screen every time it is
closed. This is because the newer functionalities require that
the hash of index tree be printed. For this reason:
- bTreeTableHashOnScreen prints table hash on screen.
- bTreeIndexTreehashOnScreen prints index tree hash on screen

Furthermore, createBTreeFromTable creates a bTree from a binary
table of entries (its argument is the name of the file).